requirements.txt
discord.py>=2.3.0
openai>=1.76.0
python-dotenv>=1.1.0
colorama>=0.4.6
aiohttp>=3.8.0